requirements.txt
bs4
python-docx
